code cleanup - refactored get appconfig methods into one
