+++ src/main/scala/uk/gov/ons/bi/dataload/linker/LinkedBusinessBuilder.scala
@@ -5,7 +5,6 @@ import org.apache.spark.rdd.RDD
 import uk.gov.ons.bi.dataload.model._
 import uk.gov.ons.bi.dataload.utils.Transformers
 
-
 object LinkedBusinessBuilder {
   // NOTE:
   // This needs to be an object, not a Singleton, because we get weird Spark "Task not serializable"
